requirements.txt
django==3.2
djangorestframework==3.14.0
psycopg2-binary==2.9.6
wheel
drf-yasg==1.21.5
Faker==18.10.1
